% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{cov.test.multi}
\alias{cov.test.multi}
\title{Test of Multiple Covariance Matrix}
\usage{
cov.test.multi(X, label, alpha = 0.05)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample size, total sample mean, within group mean, within group sample covariance, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion.
}
\description{
Tests of multiple multivariate normal overall covariance arrays when the overall mean is unknown.
}
\references{
高惠璇. 应用多元统计分析. 北京大学出版社, 2005: 88-89.
}
