% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{ind.test.multi}
\alias{ind.test.multi}
\title{Test of Independence}
\usage{
ind.test.multi(data, subdim = FALSE, alpha = 0.05)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{subdim}{The dimensions of submatrices. The default is FALSE, which means the independence of all components of the random vector will be tested.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, dimension, sample mean, sample diviation, sample diviation of submatrix, statistics, modify factor, df of Chi2, p value, critical value and conclusion.
}
\description{
Independence tests for multivariate normal populations.
}
\references{
高惠璇. 应用多元统计分析. 北京大学出版社, 2005: 92-94.
}
