% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{mean.test.multi}
\alias{mean.test.multi}
\title{Test of Multiple Mean Vectors}
\usage{
\method{mean}{test.multi}(X, label, alpha = 0.05)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, between sum of squares, statistics, df of Wilk's Lambda, df of Bartlett's Chi2, df of Rao's F, p value, approximate conclusion.
}
\description{
Test of Multiple Mean Vectors
}
