% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{mean.test.single}
\alias{mean.test.single}
\title{MNormTest}
\usage{
\method{mean}{test.single}(data, mu0, Sigma0 = FALSE, alpha = 0.05)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{mu0}{The mean vector when the null hypothesis is true.}

\item{Sigma0}{(optional) The population covariance matrix. Default is FALSE which means the covariance matrix is unknown.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample mean, sample diviation, statistics, df of T2, df of F, p value, critical value and conclusion.
}
\description{
This package provides functions to test the mean vector and covariance matrix of multivariate normal distribution.
}
\author{
Xifeng Zhang
Test of Single Mean Vector
}
