% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{mean.test.two}
\alias{mean.test.two}
\title{Test of Two Mean Vectors}
\usage{
\method{mean}{test.two}(
  data1,
  data2,
  alpha = 0.05,
  equal = TRUE,
  method = c("None", "Coupled", "Transformed")
)
}
\arguments{
\item{data1}{A matrix or data frame of group 1.}

\item{data2}{A matrix or data frame of group 2.}

\item{alpha}{The significance level. Default is 0.05.}

\item{equal}{A boolean value. Default is TRUE. If TRUE, the covariance matrix is equal. If FALSE, the covariance matrix is not equal.}

\item{method}{A string value. Default is "None". When equal is FALSE, you must choose a method in "Coupled" or "Transformed". Choose "Coupled" when the sample size of two groups is equal. Choose "Transformed" when the sample size of two groups is not equal. If you want to use the likelihood ratio test, please use the function "mean.test.multi".}
}
\value{
A list containing the hypothesis, sample mean, sample diviation, statistics, df of T2, df of F, p value, critical value and conclusion.
}
\description{
Test of Two Mean Vectors
}
