% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meancov.Test}
\alias{meancov.Test}
\title{Test of Mean and Covariance Matrix at the Same Time}
\usage{
meancov.Test(X, label, alpha = 0.05, full = FALSE)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}

\item{full}{A boolean value. Default is FALSE. If TRUE, the full output will be displayed.}
}
\value{
A list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion.
}
\description{
Simultaneous testing of multiple multivariate normal overall mean vectors and covariance matrices.
}
\examples{
data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
meancov.Test(chart, species)
meancov.Test(chart, species, full = TRUE)
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.90-91.
}
