% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{covTest.single}
\alias{covTest.single}
\title{Test of Single Covariance Matrix}
\usage{
covTest.single(data, Sigma0, ball = FALSE, alpha = 0.05)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{Sigma0}{The covariance matrix when the null hypothesis is true.}

\item{ball}{A boolean value. Default is FALSE. If FALSE, the covariance matrix is Sigma0 (known). If TRUE and the Sigma0 unit matrix, the Mauchly ball test is performed. If TRUE but Sigma0 is not a unit matrix, the covariance array is tested to see if it is sigma^2*Sigma0 (sigma^2 is unknown).}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample mean, sample diviation, statistics, df of Chi2, p value, critical value and conclusion.
}
\description{
Hypothesis testing of covariance matrices for a single normal population when the mean vector is unknown.
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.83-88.
}
