% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{covTest.single}
\alias{covTest.single}
\title{Test of Single Covariance Matrix}
\usage{
covTest.single(data, Sigma0, ball = FALSE, alpha = 0.05, detail = TRUE)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{Sigma0}{The covariance matrix when the null hypothesis is true.}

\item{ball}{A boolean value. Default is FALSE. If FALSE, the covariance matrix is Sigma0 (known). If TRUE and the Sigma0 unit matrix, the Mauchly's ball test is performed. If TRUE but Sigma0 is not a unit matrix, the covariance array is tested to see if it is sigma^2*Sigma0 (sigma^2 is unknown).}

\item{alpha}{The significance level. Default is 0.05.}

\item{detail}{A boolean value. Default is TRUE. If TRUE, the detail of the test result will be displayed.}
}
\value{
A list containing the hypothesis, sample mean, sample deviation, statistics, degree of freedom, p value, critical value and conclusion will be returned.
}
\description{
Hypothesis testing of covariance matrices for a single normal population when the mean vector is unknown.
}
\examples{
data(iris)
X <- iris[, 1:4]
covTest.single(X, diag(1, 4))
covTest.single(X, diag(1, 4), ball = TRUE)
covTest.single(X, diag(2, 4), ball = TRUE)
covTest.single(X, diag(1, 4), full = TRUE)
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.83-88.
}
\author{
Xifeng Zhang
}
