% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{indTest.multi}
\alias{indTest.multi}
\title{Test of Independence}
\usage{
indTest.multi(data, subdim = FALSE, alpha = 0.05, full = FALSE)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{subdim}{The dimensions of submatrices. The default is FALSE, which means the independence of all components of the random vector will be tested.}

\item{alpha}{The significance level. Default is 0.05.}

\item{full}{A boolean value. Default is FALSE. If TRUE, the full output will be displayed.}
}
\value{
If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample mean, sample diviation, sample diviation of submatrix, statistics, modify factor, degree of freedom and conclusion will be returned.
}
\description{
Independence tests for multivariate normal populations.
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.92-94.
}
