% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meancov.Test}
\alias{meancov.Test}
\title{Test of Mean and Covariance Matrix at the Same Time}
\usage{
meancov.Test(X, label, alpha = 0.05, detail = TRUE)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}

\item{detail}{A boolean value. Default is TRUE. If TRUE, the detail of the test result will be displayed.}
}
\value{
If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion will be returned.
}
\description{
Hypothesis testing of multiple multivariate normal population mean vectors and covariance matrices (Test simultaneously).
}
\examples{
data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
meancov.Test(chart, species)
meancov.Test(chart, species, full = TRUE)
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.90-91.
}
\author{
Xifeng Zhang
}
