% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{MeanCov.test}
\alias{MeanCov.test}
\title{Test of Mean and Covariance Matrix at the Same Time}
\usage{
MeanCov.test(X, label, alpha = 0.05)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion.
}
\description{
Simultaneous testing of multiple multivariate normal overall mean vectors and covariance matrices.
}
\references{
高惠璇. 应用多元统计分析. 北京大学出版社, 2005: 90-91.
}
