% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meanTest.multi}
\alias{meanTest.multi}
\title{Test of Multiple Mean Vectors}
\usage{
meanTest.multi(X, label, alpha = 0.05)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, between sum of squares, statistics, df of Wilk's Lambda, df of Bartlett's Chi2, df of Rao's F, p value, approximate conclusion.
}
\description{
Mean vector test for multiple multivariate normal totals when the covariance array is equal (multivariate analysis of variance).
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.80-83.
}
