% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meanTest.multi}
\alias{meanTest.multi}
\title{Test of Multiple Mean Vectors}
\usage{
meanTest.multi(X, label, alpha = 0.05, full = FALSE)
}
\arguments{
\item{X}{The data matrix which is a matrix or data frame.}

\item{label}{A vector of group labels.}

\item{alpha}{The significance level. Default is 0.05.}

\item{full}{A boolean value. Default is FALSE. If TRUE, the full output will be displayed.}
}
\value{
If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, between sum of squares, statistics, degree of freedom, p value, critical value and conclusion will be returned.
}
\description{
Mean vector test for multiple multivariate normal totals when the covariance array is equal (multivariate analysis of variance). Note that this function provides two approximations (Bartlett's chi2 and Rao's F) to compute the p-value and the critical value, and gives the realized value of Wilks Lambda statistic and its degrees of freedom (set full=TRUE to see it), if you want to do an exact test, look up Wilks Lambda according to the realized value of the statistic and its degrees of freedom statistic quantile table to solve it manually.
}
\examples{
data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
meanTest.multi(chart, species)
meanTest.multi(chart, species, full = TRUE)
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.80-83.
}
\author{
Xifeng Zhang
}
