% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meanTest.single}
\alias{meanTest.single}
\title{Test of Single Mean Vector}
\usage{
meanTest.single(data, mu0, Sigma0 = FALSE, alpha = 0.05)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{mu0}{The mean vector when the null hypothesis is true.}

\item{Sigma0}{(optional) The population covariance matrix. Default is FALSE which means the covariance matrix is unknown.}

\item{alpha}{The significance level. Default is 0.05.}
}
\value{
A list containing the hypothesis, sample mean, sample diviation, statistics, df of T2, df of F, p value, critical value and conclusion.
}
\description{
Hypothesis testing of the mean vector of a single multivariate normal population when the covariance matrix is known or unknown.
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.66-68.
}
\author{
Xifeng Zhang
}
