% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meanTest.single}
\alias{meanTest.single}
\title{Test of Single Mean Vector}
\usage{
meanTest.single(data, mu0, Sigma0 = FALSE, alpha = 0.05, full = FALSE)
}
\arguments{
\item{data}{The data matrix which is a matrix or data frame.}

\item{mu0}{The mean vector when the null hypothesis is true.}

\item{Sigma0}{(optional) The population covariance matrix. Default is FALSE which means the covariance matrix is unknown.}

\item{alpha}{The significance level. Default is 0.05.}

\item{full}{A boolean value. Default is FALSE. If TRUE, the full output will be displayed.}
}
\value{
If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample mean, sample deviation, statistics, degree of freedom, p value, critical value and conclusion will be returned.
}
\description{
Hypothesis testing of the mean vector of a single multivariate normal population when the covariance matrix is known or unknown.
}
\examples{
data(iris)
X <- iris[, 1:4]
mu0 <- c(5.8, 3.0, 4.3, 1.3)
meanTest.single(X, mu0)
meanTest.single(X, mu0, Sigma0 = diag(1, 4))
meanTest.single(X, mu0, full = TRUE)
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.66-68.
}
\author{
Xifeng Zhang
}
