% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNormTest.R
\name{meanTest.two}
\alias{meanTest.two}
\title{Test of Two Mean Vectors}
\usage{
meanTest.two(
  data1,
  data2,
  alpha = 0.05,
  equal = TRUE,
  method = c("None", "Coupled", "Transformed"),
  full = FALSE
)
}
\arguments{
\item{data1}{A matrix or data frame of group 1.}

\item{data2}{A matrix or data frame of group 2.}

\item{alpha}{The significance level. Default is 0.05.}

\item{equal}{A boolean value. Default is TRUE. If TRUE, the covariance matrix is equal. If FALSE, the covariance matrix is not equal.}

\item{method}{A string value. Default is "None". When equal is FALSE, you must choose a method in "Coupled" or "Transformed". Choose "Coupled" when the sample size of two groups is equal. Choose "Transformed" when the sample size of two groups is not equal.}

\item{full}{A boolean value. Default is FALSE. If TRUE, the full output will be displayed.}
}
\value{
If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample mean, sample deviation, statistics, degree of freedom, p value, critical value and conclusion will be returned.
}
\description{
Hypothesis testing for equality of the mean vectors of two multivariate normal populations when the covariance matrices are equal or unequal.
}
\examples{
data(iris)
X <- iris[1:50, 1:4]
Y <- iris[51:100, 1:4]
meanTest.two(X, Y)
meanTest.two(X, Y, full = TRUE)
meanTest.two(X, Y, equal = FALSE, method = "Coupled")
meanTest.two(X, Y, equal = FALSE, method = "Transformed")
}
\references{
Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.76-80.
}
\author{
Xifeng Zhang
}
